Isolated student events fetch and student event delete logic into utils file
